sm8250-common: init: Set NV carrier ID from exp_operator_switch.config value
This value (known as ro.boot.oplus_nv_id on some devices) is used to set the NV carrier ID on stock
which serve to identify the region NV ID (EUEX, RU, IN, TH, etc).

We will use that value to import a property file to set some region specific props like device, name and model.

Change-Id: I71e72f080156f73752db014734900931c3b46880